General update.

+++ trunk/tinySIP/include/tinysip/transports/tsip_transport_layer.h
@@ -59,7 +59,7 @@ int tsip_transport_layer_remove(tsip_transport_layer_t* self, const char* descri
 int tsip_transport_layer_send(const tsip_transport_layer_t* self, const char *branch, const tsip_message_t *msg);
 
 int tsip_transport_createTempSAs(const tsip_transport_layer_t *self);
-int tsip_transport_ensureTempSAs(const tsip_transport_layer_t *self, const tsip_response_t *r401_407, int32_t expires);
+int tsip_transport_ensureTempSAs(const tsip_transport_layer_t *self, const tsip_response_t *r401_407, int64_t expires);
 int tsip_transport_startSAs(const tsip_transport_layer_t* self, const void* ik, const void* ck);
 int tsip_transport_cleanupSAs(const tsip_transport_layer_t *self);
